% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghost.R
\name{is.ggghost}
\alias{is.ggghost}
\title{Reports Whether x is a ggghost Object}
\usage{
is.ggghost(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
logical; \code{TRUE} if \code{x} inherits class \code{ggghost}
}
\description{
Reports Whether x is a ggghost Object
}

