% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghost.R
\name{print.ggghost}
\alias{print.ggghost}
\title{Collect ggghost calls and produce the ggplot output}
\usage{
\method{print}{ggghost}(x, ...)
}
\arguments{
\item{x}{A ggghost object to be made into a ggplot grob}

\item{...}{Not used, provided for \code{print.default} generic consistency.}
}
\value{
The ggplot plot data (invisibly). Used for the side-effect of producing a ggplot plot.
}
\description{
Collect ggghost calls and produce the ggplot output
}
