% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghost.R
\name{reanimate}
\alias{lazarus}
\alias{reanimate}
\title{Bring a ggplot to Life (re-animate)}
\usage{
reanimate(object, gifname = "ggghost.gif", interval = 1, ani.width = 600,
  ani.height = 600)

lazarus(object, gifname = "ggghost.gif", interval = 1, ani.width = 600,
  ani.height = 600)
}
\arguments{
\item{object}{A ggghost object to animate}

\item{gifname}{Output filename to save the .gif to (not including any path,
will be saved to current directory)}

\item{interval}{A positive number to set the time interval of the animation
(unit in seconds); see \code{animation::ani.options}}

\item{ani.width}{width of image frames (unit in px); see
\code{animation::ani.options}}

\item{ani.height}{height of image frames (unit in px); see
\code{animation::ani.options}}
}
\value{
\code{TRUE} if it gets that far
}
\description{
Creates an animation showing the stepwise process of building up a ggplot.
Successively adds calls from a ggghost object and then combines these into an
animated GIF.
}
\examples{
\dontrun{
## create an animation showing the process of building up a plot
reanimate(z, "mycoolplot.gif")
}
}

