% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghost.R
\name{supp_data<-}
\alias{supp_data<-}
\title{Attach supplementary data to a ggghost object}
\usage{
supp_data(x) <- value
}
\arguments{
\item{x}{A ggghost object to which the supplementary data should be
attached}

\item{value}{Supplementary data to attach to the ggghost object, probably
used as an additional data input to a \code{scale_*}  or \code{geom_*} call}
}
\value{
The original object with \code{suppdata} attribute
}
\description{
Attach supplementary data to a ggghost object
}

