% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghost.R
\name{supp_data}
\alias{supp_data}
\title{Inspect the supplementary data attached to a ggghost object}
\usage{
supp_data(x)
}
\arguments{
\item{x}{A ggghost object}
}
\value{
A list with two elements: the name of the supplementary data, and the
  supplementary data itself
}
\description{
Inspect the supplementary data attached to a ggghost object
}

